Split cluster and linked cluster planning to separate modules
